capturer/requirements.txt
scipy
simplecv
pygame
kafka-python
avro
